# support encoding FSA with characterLen > 1