BuiltTree tracks the largest rune seen, this allows the Encoder to
choose a characterLen large enough for all characters used in the FSA.

Encoder chooses the smallest characterLen that will work, and generates
correct header.

Encoder encodes each character into the correct space dictated by the
characterLen.  The encoding is designed to be compatible with the
existing decodeCharacter method (already present in the Decoder).

Encoder tests were added to verify encoding out put of files when the
characterLen of 2 and 4 are required.

Decoder tests were added to verify correct behavior with inputs
containing characterLen 2 and 4, though the Decoder itself was not
modified.

Encoder was changed to make WriteTo the primary API, since it can
stream out words to the underlying Writer as opposed to writing them
all out in memory first.  The Encode() method is now a wrapper around
the WriteTo streaming version.

This allows a performance optimization, since the Encoder is writing
out words at a time, and the words are fixed size, we can reuse a
buffer.  This change means the Encoder generates significantly less
garbage and makes fewer calls into the GC subsystem.

A new benchmark BenchmarkEncodeTo was added to quantify this:

$ benchcmp before.txt after.txt
benchmark               old ns/op     new ns/op     delta
BenchmarkEncodeTo-4     442626        377830        -14.64%

benchmark               old allocs     new allocs     delta
BenchmarkEncodeTo-4     4968           2833           -42.98%

benchmark               old bytes     new bytes     delta
BenchmarkEncodeTo-4     144112        86688         -39.85%

NOTE: one side-effect of this change is that Encoder is not
thread-safe.  It can be reused, but should only be used by a single
goroutine at a time.