@mschoch 2017-02-28
Also, since I just realized there is another PR for an alternate solution.  I chose to keep the format fixed size, even though it wastes space, because it would more easily allow for mmap-backed runtime, without forcing the entire FSA into RAM (which is the next direction I wanted to explore).